+++ .eslintrc.json
@@ -1,4 +1,7 @@
 {
+  "env": {
+    "browser": true
+  },
   "extends": [
     "airbnb",
     "prettier"
@@ -9,6 +12,8 @@
   "rules": {
     "prettier/prettier": [
       "error"
-    ]
+    ],
+    "max-classes-per-file": 0,
+    "func-names": 0
   }
 }
