+++ .gitignore
@@ -1,3 +1,7 @@
+### Builds
+dist/main.js
+dist/main.js*
+
 ### Visual Studio Code ###
 .vscode/*
 !.vscode/tasks.json
