+++ .prettierrc
@@ -1,5 +1,6 @@
 {
   "printWidth": 120,
+  "tabWidth": 4,
   "singleQuote": true,
   "camelcase": false
 }
