--- game/index.html
@@ -1,72 +1,0 @@
-<html>
-
-  <head>
-    <link rel="stylesheet" href="style.css" />
-    <script type="text/javascript" src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
-    <script type="text/javascript" src="https://unpkg.com/@stdlib/stdlib@0.0.32/dist/stdlib-flat.min.js"></script>
-    <script type="text/javascript" src="http://cdn.jsdelivr.net/npm/jstat@latest/dist/jstat.min.js"></script>
-    <script type="text/javascript" src="../lib/util.js" type="module"></script>
-    <script type="text/javascript" src="../lib/discrete.js" type="module"></script>
-    <script type="text/javascript" src="../lib/combinatorics.js" type="module"></script>
-    <script type="text/javascript" src="game-animation.js" type="module"></script>
-    <script type="text/javascript" src="../lib/simulation.js" type="module"></script>
-    
-    <script type='text/javascript'>
-    </script>
-  </head>
-
-  <body onload="runGame(makeEnv())">
-    <div class="button start-pause start">
-      <!--div id="start" class="button">Start</div--> Start
-    </div>
-    <div id="header">
-      <div>Day: <span id="day">0</span></div>
-      <table>
-	<tr>
-	  <td></td>
-	  <td>Infected</td>
-	  <td>Deaths</td>
-	  <td>Cost of Hospitalizations</td>
-	  <td>Cost of Deaths</td>
-	  <td>Lost Economic Activity</td>
-	  <td>All Costs</td>
-	</tr>
-	<tr>
-	  <td>Today:</td>
-	  <td id="infected_current">0</td>
-	  <td id="deaths_current">0</td>
-	  <td id="cost_medical_current">$0</td>	  
-	  <td id="cost_death_current">$0</td>
-	  <td id="cost_economic_current">$0</td>
-	  <td id="cost_all_current">$0</td>
-	</tr>
-	<tr>
-	  <td>Total:</td>
-	  <td id="infected_total">0</td>
-	  <td id="deaths_total">0</td>
-	  <td id="cost_medical_total">$0</td>
-	  <td id="cost_death_total">$0</td>
-	  <td id="cost_economic_total">$0</td>
-	  <td id="cost_all_total">$0</td>
-	</tr>
-      </table>
-      
-    </div>
-
-    <div id="body">
-      <div id="close" class="button">
-	Full Lock-Down<br/>
-	<span>(Close all non-essential business; heavily restrict travel.)</span>
-      </div>
-      <div id="new-normal" class="button">
-	"New Normal"<br/>
-	<span>(Physical distancing, mandatory mask usage, recommended quarantine for travelers. In-person business slows down.)</span>
-      </div>
-      <div id="open" class="button">
-	Fully Re-Open<br/>
-	<span>Allow all businesses to re-open; no enforcement of travel, mask wearing, or distancing.</span>
-      </div>
-    </div>
-  </body>
-
-</html>
