+++ package.json
@@ -2,13 +2,11 @@
   "name": "pandemic-game",
   "version": "1.0.0",
   "description": "A simple web game / policy simulator to build intuition around the best way to respond to a pandemic",
-  "main": "game/index.html",
-  "directories": {
-    "lib": "lib"
-  },
+  "private": true,
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "lint-fix": "eslint \"./**/*.js\""
+    "dev": "webpack-dev-server --mode=development --config webpack.config.js",
+    "build": "webpack",
+    "lint-fix": "eslint \"./src/**/*.js\""
   },
   "repository": {
     "type": "git",
@@ -26,6 +24,11 @@
   },
   "homepage": "https://github.com/blakeelias/pandemic-game#readme",
   "devDependencies": {
+    "@babel/cli": "^7.12.1",
+    "@babel/core": "^7.12.3",
+    "@babel/node": "^7.12.6",
+    "@babel/preset-env": "^7.12.1",
+    "babel-loader": "^8.2.1",
     "eslint": "^7.2.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-prettier": "^6.15.0",
@@ -34,6 +37,13 @@
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-react": "^7.21.5",
     "eslint-plugin-react-hooks": "^4.0.0",
-    "prettier": "^2.1.2"
+    "prettier": "^2.1.2",
+    "webpack": "^5.4.0",
+    "webpack-cli": "^4.2.0"
+  },
+  "dependencies": {
+    "@stdlib/stdlib": "0.0.92",
+    "jquery": "^3.5.1",
+    "jstat": "^1.9.4"
   }
 }
