+++ webpack.config.js
@@ -1,0 +1,43 @@
+const path = require('path');
+
+module.exports = {
+    entry: './src/index.js',
+    output: {
+        filename: 'main.js',
+        path: path.resolve(__dirname, 'dist'),
+    },
+    module: {
+        rules: [
+            {
+                test: /\.m?js$/,
+                exclude: /node_modules/,
+                use: {
+                    loader: 'babel-loader',
+                    options: {
+                        presets: [['@babel/preset-env', { targets: 'defaults' }]],
+                    },
+                },
+            },
+        ],
+    },
+    devServer: {
+        contentBase: path.resolve(__dirname, 'dist'),
+        compress: true,
+        port: 9000,
+        open: true,
+        stats: {
+            assets: false,
+            children: false,
+            chunks: false,
+            chunkModules: false,
+            colors: true,
+            entrypoints: false,
+            hash: false,
+            modules: false,
+            timings: false,
+            version: false,
+        },
+    },
+    watch: false,
+    devtool: 'source-map',
+};
